feat: 가게 리뷰 목록 조회 api 구현

+++ src/main/java/umc/spring/domain/Review.java
@@ -29,7 +29,7 @@ public class Review extends BaseEntity {
 
     @Column(nullable = false)
     @ColumnDefault("0")
-    private Integer star;
+    private Float star;
 
     @Column(nullable = false, length = 20)
     private String title;
